Prevent the user from changing system groups (#16)

+++ app/controllers/system/groups_controller.rb
@@ -16,12 +16,14 @@ class System::GroupsController < ApplicationController
   end
 
   def update
+    return render_failure unless group.can_edit?
     group.name = params[:name]
     group.save!
     render_success
   end
 
   def destroy
+    return render_failure unless group.can_edit?
     group.destroy
     render_success
   end
